% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{apply.lubridate.fun}
\alias{apply.lubridate.fun}
\title{apply.lubridate.fun}
\usage{
apply.lubridate.fun(format, txt, tz = getOption("date.reader")$tz)
}
\arguments{
\item{format}{character; the name of the date format (case insensitive)}

\item{txt}{character; the text to convert to a date}

\item{tz}{character; the time zone}
}
\value{
A POSIXct object, or NA
}
\description{
Applies a lubridate function
}
\note{
Internal function that is not exported
}
\examples{
\dontrun{
    apply.lubridate.fun("mdy", "01/23/2017", "UTC")
  }
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

