% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.POSIXct.character}
\alias{as.POSIXct.character}
\title{as.POSIXct.character}
\usage{
\method{as.POSIXct}{character}(x, tz = getOption("date.reader")$tz, ...)
}
\arguments{
\item{x}{character; vector to convert to a POSIX date}

\item{tz}{character; optional time zone}

\item{...}{list; other optional arguments (from as.POSIXct, ignored)}
}
\value{
POSIXct vector (or NA, if a value cannot be parsed)
}
\description{
Convers character to POSIXct
}
\note{
The default timezone matches UTC like \code{lubridate} and not
  the default of \code{\link[base]{as.POSIXct}} from the base package.
}
\examples{
# -tk
 dts <- c( '20140210', '19791118', '19720329' )
 as.POSIXct( dts )

 dts <- rep( '20140210', 10 )
 as.POSIXct( dts )

 dts <- c( '14-02-10', '79-11-18' ) # FAIL(?)
 as.POSIXct( dts )
}
\seealso{
\code{\link[base]{as.POSIXct}} \cr
  \code{\link[base]{strptime}} for \code{format} strings
  \code{\link[lubridate]{parse_date_time}} for \code{orders} strings
}

