% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute.classes}
\alias{.compute.classes}
\alias{compute.classes}
\title{Internal functions for normalizing the column classes spec for read.table}
\usage{
.compute.classes(colClasses, x)
}
\arguments{
\item{colClasses}{character; spec for classes of a data table or data frame}

\item{x}{data; the data table or data frame}
}
\value{
POSIXct object representing a date (or datetime), or NA
if the input cannot be parsed.
}
\description{
Internal functions for normalizing the column classes spec for read.table

.compute.classes
}
\note{
Internal function, a wrapper for lubridate functions

If a vector input is given, the values that cannot be parsed
will be given the value NA
}
\examples{
# Named arguments
  x <- list(age="numeric", name="character", state="character")
  df <- data.frame(
   name=c("fred", "barney"),
   age=c(52, 47),
   ssn=c("123-45-6789", "012-34-5678"),
   weight=c(214, 145))
  x <- date.reader:::.compute.classes(x,df)
  # Note should be c("character", "numeric", NA, NA)

  # Unnamed arguments
  x <- list("character", "numeric")
  x <- date.reader:::.compute.classes(x,df)
  # Note should be c("character", "numeric", "character", "numeric")

  # NULL argument
  x <- NULL
  x <- date.reader:::.compute.classes(x,df)
  # Note should be c(NA, NA, NA, NA)
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

