% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{kDayWords}
\alias{kAmPm}
\alias{kAt}
\alias{kDMY}
\alias{kDMY_H}
\alias{kDMY_HM}
\alias{kDMY_HMS}
\alias{kDay}
\alias{kDayWords}
\alias{kFormats}
\alias{kH}
\alias{kHM}
\alias{kHMS}
\alias{kMDY}
\alias{kMDY_H}
\alias{kMDY_HM}
\alias{kMDY_HMS}
\alias{kMonth}
\alias{kMonth1}
\alias{kMonth2}
\alias{kMonth3}
\alias{kMonth4}
\alias{kMonth5}
\alias{kMonth6}
\alias{kMonthName}
\alias{kMonthNum}
\alias{kOptionalDayName}
\alias{kSeparator1}
\alias{kSeparator2}
\alias{kSeparator3}
\alias{kSeparator4}
\alias{kSeparator5}
\alias{kSeparator6}
\alias{kTimeDigit}
\alias{kYMD}
\alias{kYMD.alt}
\alias{kYMD_H}
\alias{kYMD_H.alt}
\alias{kYMD_HM}
\alias{kYMD_HM.alt}
\alias{kYMD_HMS}
\alias{kYMD_HMS.alt}
\alias{kYYYYMMDD}
\alias{kYYYYMMDD_HH}
\alias{kYYYYMMDD_HHMM}
\alias{kYYYYMMDD_HHMMSS}
\alias{kYear}
\alias{kYear4}
\title{#}
\format{\preformatted{ chr [1:21] "sunday" "sun" "su" "monday" "mon" "mo" "tuesday" "tue" "tu" "wednesday" "wed" "we" ...
}}
\usage{
kDayWords

kOptionalDayName

kDay

kSeparator1

kSeparator2

kMonthName

kSeparator3

kSeparator4

kYear

kYear4

kTimeDigit

kAt

kAmPm

kSeparator6

kH

kHM

kHMS

kMonthNum

kMonth

kMonth1

kMonth2

kMonth3

kDMY

kSeparator5

kMDY

kYMD.alt

kYMD_H.alt

kYMD_HM.alt

kYMD_HMS.alt

kMonth2

kMonth4

kMonth5

kMonth6

kYMD

kDMY_H

kDMY_HM

kDMY_HMS

kMDY_H

kMDY_HM

kMDY_HMS

kYMD_H

kYMD_HM

kYMD_HMS

kYYYYMMDD

kYYYYMMDD_HH

kYYYYMMDD_HHMM

kYYYYMMDD_HHMMSS

kFormats
}
\description{
#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#

#
}
\keyword{datasets}

