% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{date.reader}
\alias{date.reader}
\alias{date.reader-package}
\title{date.reader}
\description{
Recognizes and coerces dates when converting or reading data by overloading
the \code{type.convert} and \code{read.table} methods.
}
\examples{
# -tk
}

