% Generated by roxygen2 (4.0.1): do not edit by hand
\name{date.regex}
\alias{apply.lubridate.fun}
\alias{date.regex}
\title{Utilities for dealing with date regular expressions}
\usage{
apply.lubridate.fun(format, txt, tz)
}
\arguments{
\item{format}{character; the name of the date format (case insensitive)}

\item{txt}{character; the text to convert to a date}

\item{tz}{character; the time zone}
}
\value{
A POSIXct object, or NA
}
\description{
Utilities for dealing with date regular expressions

apply.lubridate.fun
}
\note{
Internal function that is not exported
}
\examples{
date.reader:::apply.lubridate.fun("mdy", "01/23/2017", "UTC")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

