% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lookup.lubridate.fun}
\alias{lookup.lubridate.fun}
\title{lookup.lubridate.fun}
\usage{
lookup.lubridate.fun(name)
}
\arguments{
\item{name}{character; the name of the function (case insensitive)}
}
\value{
the lubridate function with that name
}
\description{
lookup.lubridate.fun
}
\note{
Internal function that is not exported
}
\examples{
lookup.lubridate.fun("mdy")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

