% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lookup.regex}
\alias{lookup.regex}
\title{lookup.regex}
\usage{
lookup.regex(name)
}
\arguments{
\item{name}{character; the name given to the regex}
}
\value{
the regex with that name
}
\description{
lookup.regex
}
\note{
Internal function that is not exported
}
\examples{
lookup.regex("mdy")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

