% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse.date.aux}
\alias{parse.date.aux}
\title{parse.date.aux}
\usage{
parse.date.aux(txt, fmt)
}
\arguments{
\item{txt}{character; value to convert to a POSIX date}

\item{fmt}{character; name of the format, e.g. "MDY"}
}
\value{
POSIXct object representing a date (or date-time), or NA
if the input cannot be parsed.
}
\description{
parse.date.aux
}
\note{
Internal function, a wrapper for lubridate functions
}
\examples{
# -tk
 parse.date.aux("January 11, 2014", "MDY")
 parse.date.aux("2014/02/16", "YMD")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

