% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.parse.date}
\alias{.parse.date}
\title{.parse.date}
\usage{
.parse.date(txt, fmt)
}
\arguments{
\item{txt}{character; value to convert to a POSIX date}

\item{fmt}{character; name of the format, e.g. "MDY"}
}
\value{
POSIXct object representing a date (or date-time), or NA
if the input cannot be parsed.
}
\description{
.parse.date
}
\note{
Internal function, a wrapper for lubridate functions
}
\examples{
.parse.date("January 11, 2014", "MDY")
 .parse.date("2014/02/16", "YMD")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

