% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse.dates}
\alias{parse.dates}
\alias{parse.dates.character}
\alias{parse.dates.data.frame}
\alias{parse.dates.data.table}
\alias{parse.dates.default}
\title{parse.dates}
\usage{
parse.dates(x, ...)

\method{parse.dates}{data.frame}(x, ...)

\method{parse.dates}{default}(x, ...)

\method{parse.dates}{character}(x, ...)

\method{parse.dates}{data.table}(x, ...)
}
\arguments{
\item{x}{data; the table to fix}

\item{...}{additional arguments passed to \code{utils::read.table}}
}
\value{
data frame or data table

If a character column can be interpreted as dates, then translate
values into POSIXct objects.
}
\description{
parse.dates
}
\examples{
name <- c("bob", "fred", "sally")
  birthday <- c("01/22/1993", "02/25/1980", "03/31/1970")
  birthday <- as.factor(birthday)
  table <- data.frame(name,birthday)
  table1 <- data.table(table)
  table.new <- parse.dates(table)
  table.new1 <- parse.dates(table1)
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

