% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.table}
\alias{read.table}
\title{read.table}
\usage{
read.table(...)
}
\arguments{
\item{...}{additional arguments passed to \code{utils::read.table}}
}
\value{
data frame.

Same as \code{utils::read.table} except when a character or factor column can
be coerced to a POSIXct date.
}
\description{
read.table
}
\examples{
# -tk
}
\seealso{
\code{\link[utils]{read.table}}
  \code{\link[base]{as.POSIXct}}
}

