% Generated by roxygen2 (4.0.1): do not edit by hand
\name{string.to.POSIXct}
\alias{string.to.POSIXct}
\title{string.to.POSIXct}
\usage{
string.to.POSIXct(x, orders = NA, tz = NULL)
}
\arguments{
\item{x}{character; vector to convert to a POSIX date}

\item{orders}{character; name of the syntax for the values. If
this parameter is NA, then the function will try to guess the orders
for each value. Possible orders are:
mdy, ymd, dmy
mdy_h, ymd_h, dmy_h
mdy_hm, ymd_hm, dmy_hm
mdy_hms, ymd_hms, dmy_hms}

\item{tz}{character; optional time zone}
}
\value{
POSIXct vector (or NA, if a value cannot be parsed)
}
\description{
Convers character to POSIXct
}
\examples{
# -tk
 dts <- c( '20140210', '19791118', '19720329' )
 string.to.POSIXct( dts )

 dts <- rep( '20140210', 10 )
 string.to.POSIXct( dts )

 dts <- c( '14-02-10', '79-11-18' ) # FAIL(?)
 string.to.POSIXct( dts )

 string.to.POSIXct( dts, orders='ymd' )
}
\seealso{
\code{\link[base]{as.POSIXct}} \cr
  \code{\link[base]{strptime}} for \code{format} strings
}

