% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.format.aux}
\alias{which.format.aux}
\title{which.format.aux}
\usage{
which.format.aux(txt)
}
\arguments{
\item{txt}{character; value to convert to a POSIXct date}
}
\value{
character representing the name of the date format
}
\description{
which.format.aux
}
\note{
Internal function
}
\examples{
which.format.aux("January 11, 2014")
 which.format.aux("2014/02/16")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

