% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/which.format.R
\name{which.format}
\alias{which.format}
\title{Utilities for detecting the format of date strings}
\usage{
which.format(x, autostart = getOption("date.reader.autostart"),
  nErrors = getOption("date.reader.nErrors"))
}
\arguments{
\item{x}{character; values to convert to a POSIX date}

\item{autostart}{integer; number of strings to check before deciding
the format}

\item{nErrors}{numeric; A non-negative number:
if >= 1, the number of unparsable strings to allow
if <  1, the fraction of values that are allowed to be unparsable before
giving up.}
}
\value{
character, name for the date format, e.g. mdy, ymd_hm, etc.
}
\description{
Utilities for detecting the format of date strings

If the number of unparseable strings exceeds nErrors, return NA,
which means that the strings are either not dates, or not a consistent
format.
}
\examples{
x <- c("January 11, 2014", "February 15, 1958", "March 3, 1969")
  which.format(x, autostart=3, nErrors=1)
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

