% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.POSIXct.character}
\alias{as.POSIXct.character}
\title{as.POSIXct.character}
\usage{
\method{as.POSIXct}{character}(x, fmt = NA)
}
\arguments{
\item{x}{character; vector to convert to a POSIX date}

\item{fmt}{character; name of the format for the values. If
this parameter is NA, then the function will try to guess the format
for each value}
}
\value{
POSIXct vector (or NA, if a value cannot be parsed)
}
\description{
Convers character to POSIXct
}
\examples{
# -tk
 dts <- c( '20140210', '19791118', '19720329' )
 as.POSIXct.character( dts )
 dts <- rep( '20140210', 100 )
 dts <- c( '14-02-10', '79-11-18' ) # FAIL(?)
 as.POSIXct.character( dts )
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

