% Generated by roxygen2 (4.0.1): do not edit by hand
\name{can.posix}
\alias{can.posix}
\title{can.posix}
\usage{
can.posix(x, nTrials = min(length(x), 100), nErrors = 0)
}
\arguments{
\item{x}{character; values to parse}

\item{nTrials}{integer; number of strings to check before deciding
the format}

\item{nErrors}{integer; the number of unparsable strings to allow.
If the number of unparseable strings exceeds nErrors, return NA,
which means that the strings are either not dates, or not a consistent
format.}
}
\value{
logical. If the values can be interpreted as dates (with
the same format), it returns TRUE. It returns false if they cannot
be interpreted as dates, or if there is no consistent format.
}
\description{
can.posix
}
\examples{
can.posix("12-11-10")
  can.posix("13-13-13")
}
\seealso{
\code{\link{which.format}}
}

