% Generated by roxygen2 (4.0.2): do not edit by hand
\name{can.posix}
\alias{can.posix}
\title{can.posix}
\usage{
can.posix(x, autostart = options::get_option(date.reader$autostart,
  length(x)), nErrors = options::get_option(date.reader$nErrors, 0))
}
\arguments{
\item{x}{character; values to parse}

\item{autostart}{integer; number of strings to check before deciding
the format}

\item{nErrors}{numeric; the fraction of unparsable strings to allow.
If the fraction of unparseable strings exceeds nErrors, return NA,
which means that the strings are either not dates, or not a consistent
format.}
}
\value{
logical. If the values can be interpreted as dates (with
the same format), it returns TRUE. It returns false if they cannot
be interpreted as dates, or if there is no consistent format.
}
\description{
Determine if an object can be coerced to a POSIXct vector
}
\examples{
can.posix("12-11-10")
  can.posix("13-13-13")
}
\seealso{
\code{\link{which.orders}} \cr
  \code{canCoerce} from the methods package for similar methods
}

