% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{date.regex}
\alias{all.regex.names}
\alias{date.regex}
\alias{kAmPm}
\alias{kAt}
\alias{kDMY}
\alias{kDMY_H}
\alias{kDMY_HM}
\alias{kDMY_HMS}
\alias{kDay}
\alias{kDayWords}
\alias{kFormats}
\alias{kH}
\alias{kHM}
\alias{kHMS}
\alias{kMDY}
\alias{kMDY_H}
\alias{kMDY_HM}
\alias{kMDY_HMS}
\alias{kMonth}
\alias{kMonth1}
\alias{kMonth2}
\alias{kMonth3}
\alias{kMonth4}
\alias{kMonth5}
\alias{kMonth6}
\alias{kMonthName}
\alias{kMonthNum}
\alias{kMonthWords}
\alias{kOptionalDayName}
\alias{kSeparator1}
\alias{kSeparator2}
\alias{kSeparator3}
\alias{kSeparator4}
\alias{kSeparator5}
\alias{kSeparator6}
\alias{kTimeDigit}
\alias{kYMD}
\alias{kYMD.alt}
\alias{kYMD_H}
\alias{kYMD_H.alt}
\alias{kYMD_HM}
\alias{kYMD_HM.alt}
\alias{kYMD_HMS}
\alias{kYMD_HMS.alt}
\alias{kYYYYMMDD}
\alias{kYYYYMMDD_HH}
\alias{kYYYYMMDD_HHMM}
\alias{kYYYYMMDD_HHMMSS}
\alias{kYear}
\alias{kYear4}
\title{Constants used internally to date.reader packaged}
\format{\preformatted{ chr [1:23] "january" "jan" "february" "feb" "march" "mar" "april" "april" "may" "june" ...
}}
\usage{
\method{all}{regex.names}()

kMonthWords

kDayWords

kOptionalDayName

kDay

kSeparator1

kSeparator2

kMonthName

kSeparator3

kSeparator4

kYear

kYear4

kTimeDigit

kAt

kAmPm

kSeparator6

kH

kHM

kHMS

kMonthNum

kMonth

kMonth1

kMonth2

kMonth3

kDMY

kSeparator5

kMDY

kYMD.alt

kYMD_H.alt

kYMD_HM.alt

kYMD_HMS.alt

kMonth2

kMonth4

kMonth5

kMonth6

kYMD

kDMY_H

kDMY_HM

kDMY_HMS

kMDY_H

kMDY_HM

kMDY_HMS

kYMD_H

kYMD_HM

kYMD_HMS

kYYYYMMDD

kYYYYMMDD_HH

kYYYYMMDD_HHMM

kYYYYMMDD_HHMMSS

kFormats
}
\description{
These are the constants used internally by the \code{date.reader} package.
They are not exported.
}
\keyword{datasets}

