% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.regex.R
\name{lookup.regex}
\alias{lookup.regex}
\title{lookup.regex}
\usage{
lookup.regex(name)
}
\arguments{
\item{name}{character; the name given to the regex}
}
\value{
the regex with that name
}
\description{
find a regexp by name
}
\examples{
date.reader:::lookup.regex("mdy")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

