% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{parse.date}
\alias{.get_month_num}
\alias{.parse.date.hetero}
\alias{.parse.date.strict}
\alias{all.orders}
\alias{kMonthName}
\alias{kMonthWords}
\alias{parse.date}
\title{Utilities for parsing date strings}
\format{\preformatted{ chr [1:12] "jan" "feb" "mar" "apr" "may" "jun" "jul" "aug" "sep" "oct" "nov" "dec"
}}
\usage{
.parse.date.strict(txt, orders, tz = getOption("date.reader.tz"))

.parse.date.hetero(txt, orders = all.orders,
  tz = getOption("date.reader.tz"))

kMonthWords

kMonthName

.get_month_num(txtstr)

all.orders
}
\arguments{
\item{txt}{character; value(s) to convert to a POSIX date}

\item{orders}{character; name of the order, e.g. "mdy"
orders are described in
  \code{\link[lubridate]{parse_date_time}}. Each order string is
  series of formatting characters as listed \code{\link[base]{strptime}} but
  without the "%" prefix, for example "ymd" will match all the possible
  dates in year, month, day order. Formatting orders might include
  arbitrary separators. These are discarded. See details for implemented
  formats.}

\item{tz}{character; optional time zone}

\item{txtstr}{character; the string to test}
}
\value{
POSIXct object representing a date (or datetime), or NA
if the input cannot be parsed.

POSIXct object representing a date (or datetime), or NA
if the input cannot be parsed.

numeric; the month number, or -1 if none
}
\description{
Utilities for parsing date strings

.parse.date.strict

.parse.date.hetero
parses a vector of date strings. May be different formats

.get_month_num
If a string contains a fragment of the name of a month,
it returns the month number.
}
\note{
Internal function, a wrapper for lubridate functions

If a vector input is given, the values that cannot be parsed
will be given the value NA

Internal function, a wrapper for lubridate functions

"orders" argument here is a vector of orders to choose from

Internal function, not exported
}
\examples{
date.reader:::.parse.date.strict( "January 11, 2014", "mdy" )
   date.reader:::.parse.date.strict( "January 11, 2014", "MDY" )

   date.reader:::.parse.date.strict( "2014/02/16", "YMD" )
   date.reader:::.parse.date.strict( "2014-08-05", "YMD" )
   date.reader:::.parse.date.strict( c("2014-08-05","2014-08-06" ), "YMD" )
   date.reader:::.parse.date.strict( "20140805", "ymd")
   date.reader:::.parse.date.strict( "14/08/05", "ymd")
   date.reader:::.parse.date.strict( "2/2/15", "dmy" )
date.reader:::.parse.date.hetero(
     c("January 11, 2014", "2014/02/16", "2014-08-05", "20140805", "14/08/05")
   )
date.reader:::.get_month_num("blahblahjanuaryblahblah")
}
\seealso{
\code{\link[base]{as.POSIXct}}

\code{\link[base]{as.POSIXct}}
}
\keyword{datasets}

