% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{.parse.date}
\alias{.parse.date}
\title{.parse.date}
\usage{
.parse.date(txt, format, tz = NULL, check.regex = TRUE)
}
\arguments{
\item{txt}{character; value to convert to a POSIX date}

\item{format}{character; name of the format, e.g. "MDY"}

\item{tz}{character; optional time zone}

\item{check.regex}{Logical; if TRUE, check syntax of
txt before calling lubridate function.}
}
\value{
POSIXct object representing a date (or date-time), or NA
if the input cannot be parsed.
}
\description{
.parse.date
}
\note{
Internal function, a wrapper for lubridate functions
}
\examples{
\dontrun{
   .parse.date( "January 11, 2014", "MDY" )
   .parse.date( "2014/02/16", "YMD" )
   .parse.date( "2014-08-05", "YMD" )
   .parse.date( "20140805", "ymd.numeric")
   .parse.date( "14/08/05", "ymd.alt")
  }
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

