% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.parse.date}
\alias{.parse.date}
\title{.parse.date}
\usage{
.parse.date(txt, orders, tz = getOption("date.reader")$tz,
  check.regex = TRUE)
}
\arguments{
\item{txt}{character; value to convert to a POSIX date}

\item{orders}{character; name of the format, e.g. "MDY"}

\item{tz}{character; optional time zone}

\item{check.regex}{Logical; if TRUE, check syntax of \code{txt} before
  calling lubridate function.}
}
\value{
POSIXct object representing a date (or datetime), or NA
if the input cannot be parsed.
}
\description{
.parse.date
}
\note{
Internal function, a wrapper for lubridate functions
}
\examples{
date.reader:::.parse.date( "January 11, 2014", "MDY" )
   date.reader:::.parse.date( "2014/02/16", "YMD" )
   date.reader:::.parse.date( "2014-08-05", "YMD" )
   date.reader:::.parse.date( "20140805", "ymd.numeric")
   date.reader:::.parse.date( "14/08/05", "ymd.alt")
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

