% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse.dates}
\alias{parse.dates}
\title{fix.date.columns}
\usage{
parse.dates(table, ...)
}
\arguments{
\item{table}{date.frame; the table to fix}

\item{...}{additional arguments passed to \code{utils::read.table}}
}
\value{
data frame.

If a character column can be interpreted as dates, then translate
values into POSIXct objects.
}
\description{
fix.date.columns
}
\examples{
name <- c("bob", "fred", "sally")
  birthday <- c("01/22/1993", "02/25/1980", "03/31/1970")
  birthday <- as.factor(birthday)
  table <- data.frame(name,birthday)

  table.new <- parse.dates(table)
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

