% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.table}
\alias{read.table}
\title{read.table}
\usage{
read.table(file, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{utils::read.table}}
}
\value{
data frame.

If a character column can be interpreted as dates, then it translates the
value into POSIXct objects.  Otherwise, it acts as utils::read.table
}
\description{
read.table
}
\examples{
# -tk
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

