% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.table}
\alias{read.table}
\title{read.table}
\usage{
read.table(file, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{utils::read.table}}
}
\value{
data frame.

If a character column can be interpreted as dates, then it translates the
value into POSIXct objects.  Otherwise, it acts as utils::read.table
}
\description{
read.table
}
\examples{
# -tk
name <- c("bob", "fred", "sally")
birthday <- c("01/22/1993", "02/25/1980", "03/31/1970")
birthday <- as.factor(birthday)
table <- data.table(name,birthday)
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

