% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.table}
\alias{read.table}
\title{read.table}
\usage{
read.table(...)

read.table(...)
}
\arguments{
\item{file}{character; path to a file to be read in}

\item{file}{character; path to a file to be read in}
}
\value{
data frame. If a column can be interpreted as
dates, then it translates the value into POSIXct objects.
Otherwise, it acts as utils::read.table

data frame. If a column can be interpreted as
dates, then it translates the value into POSIXct objects.
Otherwise, it acts as utils::read.table
}
\description{
read.table

read.table
}
\seealso{
\code{\link[base]{as.POSIX}}

\code{\link[base]{as.POSIX}}
}

