% Generated by roxygen2 (4.0.1): do not edit by hand
\name{type.convert}
\alias{type.convert}
\title{type.convert}
\usage{
type.convert(x, ...)
}
\arguments{
\item{x}{vector of character; values to parse}
}
\value{
vector of parsed values. If the values
can be interpreted as dates, it returns a vector
of POSIXct objects. Otherwise, it calls utils::type.convert
}
\description{
type.convert
}
\seealso{
\code{\link[base]{as.POSIX}}
}

