% Generated by roxygen2 (4.0.1): do not edit by hand
\name{type.convert}
\alias{type.convert}
\title{type.convert}
\usage{
type.convert(x, ...)
}
\arguments{
\item{x}{character;}

\item{...}{list; extra arguments (from utils::type.convert, not used)}
}
\value{
a vector of parsed values. If values
can be interpreted as dates, returns a POSIXct vector otherwise, the values
from utils::type.convert
}
\description{
type.convert
}
\examples{
x <- type.convert( "2014-08-05" )
  x <- type.convert( c("2014-08-05", "2014-08-12") )
  x

  class(x)  # "POSIXct" "POSIXt"
}
\seealso{
\code{\link[base]{as.POSIXct}}
}

