% Generated by roxygen2 (4.0.1): do not edit by hand
\name{type.convert}
\alias{type.convert}
\title{type.convert}
\usage{
type.convert(x, ...)
}
\arguments{
\item{x}{character;}
}
\value{
a vector of parsed values. If values
can be interpreted as dates, returns a POSIXct vector otherwise, the values
from utils::type.convert
}
\description{
type.convert
}
\examples{
x <- type.convert( c("2014-08-05" ) )
  x
  class(x)  # "POSIXct" "POSIXt"
}
\seealso{
\code{\link[base]{as.POSIX}}
}

