% Generated by roxygen2 (4.0.2): do not edit by hand
\name{type.convert}
\alias{type.convert}
\title{type.convert}
\usage{
type.convert(x, ...)
}
\arguments{
\item{x}{character; vector to be coerces to POSIXct if possible}

\item{...}{list; extra arguments (from utils::type.convert, not used)}
}
\value{
a vector of parsed values. If values can be interpreted as dates,
returns a POSIXct vector otherwise, the values returned from
\code{utils::type.convert}
}
\description{
Coerces a character vector to POSIXct if possible. An extension to
\code{utils::type.convert}.
}
\details{
If the input has a recognized date format, then it is parsed and converted
to POSIXct. Otherwise, the input is passed to
\code{utils::type.convert}.
}
\note{
this function will only attempt to convert a vector to POSIXct if every
item in the vector has the same format.
}
\examples{
type.convert( "2014-08-05" )
  type.convert( c("2014-08-05", "2014-08-12") )
}
\seealso{
\code{\link[base]{as.POSIXct}}
  \code{\link[utils]{type.convert}}
}

