% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.format}
\alias{which.format}
\title{which.format}
\usage{
which.format(x, nTrials = 1, nErrors = 0)
}
\arguments{
\item{x}{vector of characters; values to convert to a POSIX date}

\item{nTrials}{integer; number of strings to check before deciding
the format}

\item{nErrors}{integer; the number of unparsable strings to allow.
If the number of unparseable strings exceeds nErrors, return NA,
which means that the strings are either not dates, or not a consistent
format.}
}
\value{
character representing the name of the date format
}
\description{
which.format
}
\note{
Internal function
}
\examples{
# -tk
 x <- c("January 11, 2014", "February 15, 1958", "2015/03/23")
 which.format(x, nTrials=3, nErrors=1)
}
\seealso{
\code{\link[base]{as.POSIX}}
}

