% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.format}
\alias{.which.format}
\alias{which.format}
\title{Utilities for detecting the format of date strings}
\usage{
which.format(x, nTrials = 1, nErrors = 0)

.which.format(txt)
}
\arguments{
\item{x}{character; values to convert to a POSIX date}

\item{nTrials}{integer; number of strings to check before deciding
the format}

\item{nErrors}{integer; the number of unparsable strings to allow.
If the number of unparseable strings exceeds nErrors, return NA,
which means that the strings are either not dates, or not a consistent
format.}

\item{txt}{character; value to convert to a POSIXct date}
}
\value{
character, name for the date format

character representing the name of the date format
}
\description{
Utilities for detecting the format of date strings

which.format

.which.format
}
\note{
Internal function that is not exported
}
\examples{
x <- c("January 11, 2014", "February 15, 1958", "2015/03/23")
 which.format(x, nTrials=3, nErrors=1)
\dontrun{
   date.reader:::.which.format("January 11, 2014")
   date.reader:::.which.format("2014/02/16")
 }
}
\seealso{
\code{\link[base]{as.POSIXct}}

\code{\link[base]{as.POSIXct}}
}

